worker/requirements.txt
psycopg2-binary==2.9.6
primefac==2.0.12